LMSAdaptiveFiltering Calls mylms - part2run calls part2 which calls mylms
LMS adaptive filtering is where I import a source of ambient noise in the room, convolve it with an excitation signal of random white noise. Next, I use an external source of ambient noise and convolve it with an audio signal - an adaptive filter is used to study the impulse reponse of using a normalised LMS filter, and the corresponding error. This can be tweaked and applied with the right hardware in hospital rooms.